Redesign map overlay sheet as unified list without tabs
Replace the tabbed MainTabView with a single unified list containing two
sections (Locations and Nearby). Move Notifications and More into the top
trailing toolbar as sheet presentations with close buttons. Merge bus stop
names directly into list rows instead of section headers. Add persistent
bottom toolbar with search field, MRT Map and Fare Calculator buttons.
Remove presentation background from the sheet. Replace ListHintOverlay
with ContentUnavailableView for empty states.
