+++ Buses/Classes/NavigationManager.swift
@@ -9,40 +9,14 @@ import Foundation
 
 class NavigationManager: ObservableObject {
 
-    @Published var nearbyTabPath: [ViewPath] = []
-    @Published var favoritesTabPath: [ViewPath] = []
-    @Published var notificationsTabPath: [ViewPath] = []
-    @Published var directoryTabPath: [ViewPath] = []
-    @Published var moreTabPath: [ViewPath] = []
+    @Published var mainPath: [ViewPath] = []
 
-    func popToRoot(for tab: TabType) {
-        switch tab {
-        case .nearby:
-            nearbyTabPath.removeAll()
-        case .favorites:
-            favoritesTabPath.removeAll()
-        case .notifications:
-            notificationsTabPath.removeAll()
-        case .directory:
-            directoryTabPath.removeAll()
-        case .more:
-            moreTabPath.removeAll()
-        }
+    func popToRoot() {
+        mainPath.removeAll()
     }
 
-    func push(_ viewPath: ViewPath, for tab: TabType) {
-        switch tab {
-        case .nearby:
-            nearbyTabPath.append(viewPath)
-        case .favorites:
-            favoritesTabPath.append(viewPath)
-        case .notifications:
-            notificationsTabPath.append(viewPath)
-        case .directory:
-            directoryTabPath.append(viewPath)
-        case .more:
-            moreTabPath.append(viewPath)
-        }
+    func push(_ viewPath: ViewPath) {
+        mainPath.append(viewPath)
     }
 
 }
