+++ Buses/Views/Notifications/NotificationsView.swift
@@ -9,28 +9,24 @@ import SwiftUI
 
 struct NotificationsView: View {
 
-    @EnvironmentObject var navigationManager: NavigationManager
+    @Environment(\.dismiss) var dismiss
 
     @State var notificationRequests: [UNNotificationRequest] = []
 
     var body: some View {
-        NavigationStack(path: $navigationManager.notificationsTabPath) {
+        NavigationStack {
             List(notificationRequests, id: \.identifier) { request in
                 if let busService = request.content.userInfo["busService"] as? String,
                    let stopCode = request.content.userInfo["stopCode"] as? String,
                    let stopDescription = request.content.userInfo["stopDescription"] as? String {
-                    NavigationLink(value: ViewPath.busServiceNamed(busService,
-                                                                   atLocation: stopDescription,
-                                                                   forBusStopCode: stopCode)) {
-                        HStack(alignment: .center, spacing: 16.0) {
-                            Image(.listIconBus)
-                            VStack(alignment: .leading) {
-                                Text(request.content.title)
-                                    .font(.body)
-                                    .fontWeight(.bold)
-                                Text(request.content.body)
-                                    .font(.body)
-                            }
+                    HStack(alignment: .center, spacing: 16.0) {
+                        Image(.listIconBus)
+                        VStack(alignment: .leading) {
+                            Text(request.content.title)
+                                .font(.body)
+                                .fontWeight(.bold)
+                            Text(request.content.body)
+                                .font(.body)
                         }
                     }
                     .swipeActions {
@@ -44,20 +40,6 @@ struct NotificationsView: View {
             }
             .listStyle(.insetGrouped)
             .listSectionSpacing(0)
-            .navigationDestination(for: ViewPath.self, destination: { viewPath in
-                switch viewPath {
-                case .busServiceNamed(let serviceNumber, let locationName, let busStopCode):
-                    BusServiceView(busService: BusService(serviceNo: serviceNumber,
-                                                          operator: .unknown),
-                                   locationName: locationName,
-                                   busStopCode: busStopCode,
-                                   showsAddToLocationButton: false)
-                case .busStop(let busStop):
-                    BusStopView(busStop: busStop)
-                default:
-                    Color.clear
-                }
-            })
             .refreshable {
                 reloadNotificationRequests()
             }
@@ -66,11 +48,27 @@ struct NotificationsView: View {
             }
             .overlay {
                 if notificationRequests.count == 0 {
-                    ListHintOverlay(image: "info.circle.fill", text: "Notifications.ArrivalAlerts.Hint")
+                    ContentUnavailableView {
+                        Label("Notifications.ArrivalAlerts",
+                              systemImage: "info.circle.fill")
+                    } description: {
+                        Text("Notifications.ArrivalAlerts.Hint")
+                    }
                 }
             }
             .navigationTitle("ViewTitle.Notifications")
             .navigationBarTitleDisplayMode(.inline)
+            .toolbar {
+                ToolbarItem(placement: .navigationBarTrailing) {
+                    Button {
+                        dismiss()
+                    } label: {
+                        Image(systemName: "xmark.circle.fill")
+                            .symbolRenderingMode(.hierarchical)
+                            .foregroundStyle(.secondary)
+                    }
+                }
+            }
         }
     }
